+++ giverny/src/giverny/turbulence_dataset.py
@@ -39,7 +39,7 @@ from giverny.turbulence_gizmos.basic_gizmos import *
 
 class turb_dataset():
     def __init__(self, dataset_title = '', output_path = '', auth_token = '', rewrite_interpolation_metadata = False,
-                 json_url = '/home/idies/workspace/turbulence-ceph/sciserver-turbulence/jhtdb_metadata/jhtdb_configs/jhtdb-config.json'):
+                 json_url = '/home/idies/workspace/turbulence-ceph/sciserver-turbulence/jhtdb_metadata/jhtdb_configs/jhtdb_config.json'):
         """
         initialize the class.
         """